move test utilities to utils module
