+++ corehq/apps/userreports/tests/test_pillow.py
@@ -8,7 +8,7 @@ from casexml.apps.case.models import CommCareCase
 from corehq.apps.userreports.pillow import ConfigurableIndicatorPillow
 from corehq.apps.userreports.sql import IndicatorSqlAdapter
 from corehq.apps.userreports.tasks import rebuild_indicators
-from corehq.apps.userreports.tests import get_sample_doc_and_indicators, get_sample_data_source
+from corehq.apps.userreports.tests.utils import get_sample_data_source, get_sample_doc_and_indicators
 from corehq.db import connection_manager, DEFAULT_ENGINE_ID
